Refactor: 🥝알림 및 CursorResponse totalCount 필드 타입 변경 - CursorResponse 및 관련 서비스 내 totalCount 필드의 타입을 Long으로 수정. - NotificationRepository 메서드 반환 타입 변경. - FollowService 및 DirectMessageService에서 totalCount 값을 null로 설정.

+++ src/main/java/com/codeit/otboo/domain/notification/repository/NotificationRepository.java
@@ -14,7 +14,7 @@ import org.springframework.data.repository.query.Param;
 
 public interface NotificationRepository extends JpaRepository<Notification, UUID> {
 
-    int countNotificationByReceiverId(UUID receiverId);
+    Long countNotificationByReceiverId(UUID receiverId);
     Optional<Notification> findByIdAndReceiverId(UUID id, UUID receiverId);
 
     @Query("""
